@@ -37,11 +37,11 @@ type Annotations struct {
 
 // Audio provided to or from an LLM.
 type AudioContent struct {
-	Meta        map[string]interface{} `json:"_meta,omitempty"`
-	Annotations *Annotations           `json:"annotations,omitempty"`
-	Data        []byte                 `json:"data"`
-	MIMEType    string                 `json:"mimeType"`
-	Type        string                 `json:"type"`
+	Meta        map[string]any `json:"_meta,omitempty"`
+	Annotations *Annotations   `json:"annotations,omitempty"`
+	Data        []byte         `json:"data"`
+	MIMEType    string         `json:"mimeType"`
+	Type        string         `json:"type"`
 }
 
 // Base interface for metadata with name (identifier) and title (display name) properties.
@@ -51,10 +51,10 @@ type BaseMetadata struct {
 }
 
 type BlobResourceContents struct {
-	Meta     map[string]interface{} `json:"_meta,omitempty"`
-	Blob     []byte                 `json:"blob"`
-	MIMEType *string                `json:"mimeType,omitempty"`
-	URI      string                 `json:"uri"`
+	Meta     map[string]any `json:"_meta,omitempty"`
+	Blob     []byte         `json:"blob"`
+	MIMEType *string        `json:"mimeType,omitempty"`
+	URI      string         `json:"uri"`
 }
 
 type BooleanSchema struct {
@@ -66,16 +66,16 @@ type BooleanSchema struct {
 
 // Used by the client to invoke a tool provided by the server.
 type CallToolRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // The server's response to a tool call.
 type CallToolResult struct {
-	Meta              map[string]interface{} `json:"_meta,omitempty"`
-	Content           []ContentBlock         `json:"content"`
-	IsError           *bool                  `json:"isError,omitempty"`
-	StructuredContent map[string]interface{} `json:"structuredContent,omitempty"`
+	Meta              map[string]any `json:"_meta,omitempty"`
+	Content           []ContentBlock `json:"content"`
+	IsError           *bool          `json:"isError,omitempty"`
+	StructuredContent map[string]any `json:"structuredContent,omitempty"`
 }
 
 // This notification can be sent by either side to indicate that it is cancelling a previously-issued request.
@@ -86,51 +86,51 @@ type CallToolResult struct {
 //
 // A client MUST NOT attempt to cancel its `initialize` request.
 type CancelledNotification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // Capabilities a client may support. Known capabilities are defined here, in this schema, but this is not a closed set: any client can define its own, additional capabilities.
 type ClientCapabilities struct {
-	Elicitation  map[string]interface{}            `json:"elicitation,omitempty"`
-	Experimental map[string]map[string]interface{} `json:"experimental,omitempty"`
-	Roots        map[string]interface{}            `json:"roots,omitempty"`
-	Sampling     map[string]interface{}            `json:"sampling,omitempty"`
+	Elicitation  map[string]any            `json:"elicitation,omitempty"`
+	Experimental map[string]map[string]any `json:"experimental,omitempty"`
+	Roots        map[string]any            `json:"roots,omitempty"`
+	Sampling     map[string]any            `json:"sampling,omitempty"`
 }
 
-type ClientNotification interface{}
+type ClientNotification any
 
-type ClientRequest interface{}
+type ClientRequest any
 
-type ClientResult interface{}
+type ClientResult any
 
 // A request from the client to the server, to ask for completion options.
 type CompleteRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // The server's response to a completion/complete request
 type CompleteResult struct {
-	Meta       map[string]interface{} `json:"_meta,omitempty"`
-	Completion map[string]interface{} `json:"completion"`
+	Meta       map[string]any `json:"_meta,omitempty"`
+	Completion map[string]any `json:"completion"`
 }
 
-type ContentBlock interface{}
+type ContentBlock any
 
 // A request from the server to sample an LLM via the client. The client has full discretion over which model to select. The client should also inform the user before beginning sampling, to allow them to inspect the request (human in the loop) and decide whether to approve it.
 type CreateMessageRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // The client's response to a sampling/create_message request from the server. The client should inform the user before returning the sampled message, to allow them to inspect the response (human in the loop) and decide whether to allow the server to see it.
 type CreateMessageResult struct {
-	Meta       map[string]interface{} `json:"_meta,omitempty"`
-	Content    interface{}            `json:"content"`
-	Model      string                 `json:"model"`
-	Role       Role                   `json:"role"`
-	StopReason *string                `json:"stopReason,omitempty"`
+	Meta       map[string]any `json:"_meta,omitempty"`
+	Content    any            `json:"content"`
+	Model      string         `json:"model"`
+	Role       Role           `json:"role"`
+	StopReason *string        `json:"stopReason,omitempty"`
 }
 
 // An opaque token used to represent a cursor for pagination.
@@ -139,15 +139,15 @@ type Cursor struct {
 
 // A request from the server to elicit additional information from the user via the client.
 type ElicitRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // The client's response to an elicitation request.
 type ElicitResult struct {
-	Meta    map[string]interface{} `json:"_meta,omitempty"`
-	Action  string                 `json:"action"`
-	Content map[string]interface{} `json:"content,omitempty"`
+	Meta    map[string]any `json:"_meta,omitempty"`
+	Action  string         `json:"action"`
+	Content map[string]any `json:"content,omitempty"`
 }
 
 // The contents of a resource, embedded into a prompt or tool call result.
@@ -155,10 +155,10 @@ type ElicitResult struct {
 // It is up to the client how best to render embedded resources for the benefit
 // of the LLM and/or the user.
 type EmbeddedResource struct {
-	Meta        map[string]interface{} `json:"_meta,omitempty"`
-	Annotations *Annotations           `json:"annotations,omitempty"`
-	Resource    interface{}            `json:"resource"`
-	Type        string                 `json:"type"`
+	Meta        map[string]any `json:"_meta,omitempty"`
+	Annotations *Annotations   `json:"annotations,omitempty"`
+	Resource    any            `json:"resource"`
+	Type        string         `json:"type"`
 }
 
 type EmptyResult struct {
@@ -174,24 +174,24 @@ type EnumSchema struct {
 
 // Used by the client to get a prompt provided by the server.
 type GetPromptRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // The server's response to a prompts/get request from the client.
 type GetPromptResult struct {
-	Meta        map[string]interface{} `json:"_meta,omitempty"`
-	Description *string                `json:"description,omitempty"`
-	Messages    []PromptMessage        `json:"messages"`
+	Meta        map[string]any  `json:"_meta,omitempty"`
+	Description *string         `json:"description,omitempty"`
+	Messages    []PromptMessage `json:"messages"`
 }
 
 // An image provided to or from an LLM.
 type ImageContent struct {
-	Meta        map[string]interface{} `json:"_meta,omitempty"`
-	Annotations *Annotations           `json:"annotations,omitempty"`
-	Data        []byte                 `json:"data"`
-	MIMEType    string                 `json:"mimeType"`
-	Type        string                 `json:"type"`
+	Meta        map[string]any `json:"_meta,omitempty"`
+	Annotations *Annotations   `json:"annotations,omitempty"`
+	Data        []byte         `json:"data"`
+	MIMEType    string         `json:"mimeType"`
+	Type        string         `json:"type"`
 }
 
 // Describes the name and version of an MCP implementation, with an optional title for UI representation.
@@ -203,48 +203,48 @@ type Implementation struct {
 
 // This request is sent from the client to the server when it first connects, asking it to begin initialization.
 type InitializeRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // After receiving an initialize request from the client, the server sends this response.
 type InitializeResult struct {
-	Meta            map[string]interface{} `json:"_meta,omitempty"`
-	Capabilities    ServerCapabilities     `json:"capabilities"`
-	Instructions    *string                `json:"instructions,omitempty"`
-	ProtocolVersion string                 `json:"protocolVersion"`
-	ServerInfo      Implementation         `json:"serverInfo"`
+	Meta            map[string]any     `json:"_meta,omitempty"`
+	Capabilities    ServerCapabilities `json:"capabilities"`
+	Instructions    *string            `json:"instructions,omitempty"`
+	ProtocolVersion string             `json:"protocolVersion"`
+	ServerInfo      Implementation     `json:"serverInfo"`
 }
 
 // This notification is sent from the client to the server after initialization has finished.
 type InitializedNotification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // A response to a request that indicates an error occurred.
 type JSONRPCError struct {
-	Error   map[string]interface{} `json:"error"`
-	ID      RequestId              `json:"id"`
-	JSONRPC string                 `json:"jsonrpc"`
+	Error   map[string]any `json:"error"`
+	ID      RequestId      `json:"id"`
+	JSONRPC string         `json:"jsonrpc"`
 }
 
 // Refers to any valid JSON-RPC object that can be decoded off the wire, or encoded to be sent.
-type JSONRPCMessage interface{}
+type JSONRPCMessage any
 
 // A notification which does not expect a response.
 type JSONRPCNotification struct {
-	JSONRPC string                 `json:"jsonrpc"`
-	Method  string                 `json:"method"`
-	Params  map[string]interface{} `json:"params,omitempty"`
+	JSONRPC string         `json:"jsonrpc"`
+	Method  string         `json:"method"`
+	Params  map[string]any `json:"params,omitempty"`
 }
 
 // A request that expects a response.
 type JSONRPCRequest struct {
-	ID      RequestId              `json:"id"`
-	JSONRPC string                 `json:"jsonrpc"`
-	Method  string                 `json:"method"`
-	Params  map[string]interface{} `json:"params,omitempty"`
+	ID      RequestId      `json:"id"`
+	JSONRPC string         `json:"jsonrpc"`
+	Method  string         `json:"method"`
+	Params  map[string]any `json:"params,omitempty"`
 }
 
 // A successful (non-error) response to a request.
@@ -256,41 +256,41 @@ type JSONRPCResponse struct {
 
 // Sent from the client to request a list of prompts and prompt templates the server has.
 type ListPromptsRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // The server's response to a prompts/list request from the client.
 type ListPromptsResult struct {
-	Meta       map[string]interface{} `json:"_meta,omitempty"`
-	NextCursor *string                `json:"nextCursor,omitempty"`
-	Prompts    []Prompt               `json:"prompts"`
+	Meta       map[string]any `json:"_meta,omitempty"`
+	NextCursor *string        `json:"nextCursor,omitempty"`
+	Prompts    []Prompt       `json:"prompts"`
 }
 
 // Sent from the client to request a list of resource templates the server has.
 type ListResourceTemplatesRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // The server's response to a resources/templates/list request from the client.
 type ListResourceTemplatesResult struct {
-	Meta              map[string]interface{} `json:"_meta,omitempty"`
-	NextCursor        *string                `json:"nextCursor,omitempty"`
-	ResourceTemplates []ResourceTemplate     `json:"resourceTemplates"`
+	Meta              map[string]any     `json:"_meta,omitempty"`
+	NextCursor        *string            `json:"nextCursor,omitempty"`
+	ResourceTemplates []ResourceTemplate `json:"resourceTemplates"`
 }
 
 // Sent from the client to request a list of resources the server has.
 type ListResourcesRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // The server's response to a resources/list request from the client.
 type ListResourcesResult struct {
-	Meta       map[string]interface{} `json:"_meta,omitempty"`
-	NextCursor *string                `json:"nextCursor,omitempty"`
-	Resources  []Resource             `json:"resources"`
+	Meta       map[string]any `json:"_meta,omitempty"`
+	NextCursor *string        `json:"nextCursor,omitempty"`
+	Resources  []Resource     `json:"resources"`
 }
 
 // Sent from the server to request a list of root URIs from the client. Roots allow
@@ -301,35 +301,35 @@ type ListResourcesResult struct {
 // This request is typically used when the server needs to understand the file system
 // structure or access specific locations that the client has permission to read from.
 type ListRootsRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // The client's response to a roots/list request from the server.
 // This result contains an array of Root objects, each representing a root directory
 // or file that the server can operate on.
 type ListRootsResult struct {
-	Meta  map[string]interface{} `json:"_meta,omitempty"`
-	Roots []Root                 `json:"roots"`
+	Meta  map[string]any `json:"_meta,omitempty"`
+	Roots []Root         `json:"roots"`
 }
 
 // Sent from the client to request a list of tools the server has.
 type ListToolsRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // The server's response to a tools/list request from the client.
 type ListToolsResult struct {
-	Meta       map[string]interface{} `json:"_meta,omitempty"`
-	NextCursor *string                `json:"nextCursor,omitempty"`
-	Tools      []Tool                 `json:"tools"`
+	Meta       map[string]any `json:"_meta,omitempty"`
+	NextCursor *string        `json:"nextCursor,omitempty"`
+	Tools      []Tool         `json:"tools"`
 }
 
 // Notification of a log message passed from server to client. If no logging/setLevel request has been sent from the client, the server MAY decide which messages to send automatically.
 type LoggingMessageNotification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // Hints to use for model selection.
@@ -359,8 +359,8 @@ type ModelPreferences struct {
 }
 
 type Notification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 type NumberSchema struct {
@@ -372,29 +372,29 @@ type NumberSchema struct {
 }
 
 type PaginatedRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 type PaginatedResult struct {
-	Meta       map[string]interface{} `json:"_meta,omitempty"`
-	NextCursor *string                `json:"nextCursor,omitempty"`
+	Meta       map[string]any `json:"_meta,omitempty"`
+	NextCursor *string        `json:"nextCursor,omitempty"`
 }
 
 // A ping, issued by either the server or the client, to check that the other party is still alive. The receiver must promptly respond, or else may be disconnected.
 type PingRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // Restricted schema definitions that only allow primitive types
 // without nested objects or arrays.
-type PrimitiveSchemaDefinition interface{}
+type PrimitiveSchemaDefinition any
 
 // An out-of-band notification used to inform the receiver of a progress update for a long-running request.
 type ProgressNotification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // A progress token, used to associate progress notifications with the original request.
@@ -403,11 +403,11 @@ type ProgressToken struct {
 
 // A prompt or prompt template that the server offers.
 type Prompt struct {
-	Meta        map[string]interface{} `json:"_meta,omitempty"`
-	Arguments   []PromptArgument       `json:"arguments,omitempty"`
-	Description *string                `json:"description,omitempty"`
-	Name        string                 `json:"name"`
-	Title       *string                `json:"title,omitempty"`
+	Meta        map[string]any   `json:"_meta,omitempty"`
+	Arguments   []PromptArgument `json:"arguments,omitempty"`
+	Description *string          `json:"description,omitempty"`
+	Name        string           `json:"name"`
+	Title       *string          `json:"title,omitempty"`
 }
 
 // Describes an argument that a prompt can accept.
@@ -420,8 +420,8 @@ type PromptArgument struct {
 
 // An optional notification from the server to the client, informing it that the list of prompts it offers has changed. This may be issued by servers without any previous subscription from the client.
 type PromptListChangedNotification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // Describes a message returned as part of a prompt.
@@ -442,19 +442,19 @@ type PromptReference struct {
 
 // Sent from the client to the server, to read a specific resource URI.
 type ReadResourceRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // The server's response to a resources/read request from the client.
 type ReadResourceResult struct {
-	Meta     map[string]interface{} `json:"_meta,omitempty"`
-	Contents []interface{}          `json:"contents"`
+	Meta     map[string]any `json:"_meta,omitempty"`
+	Contents []any          `json:"contents"`
 }
 
 type Request struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // A uniquely identifying ID for a request in JSON-RPC.
@@ -463,53 +463,53 @@ type RequestId struct {
 
 // A known resource that the server is capable of reading.
 type Resource struct {
-	Meta        map[string]interface{} `json:"_meta,omitempty"`
-	Annotations *Annotations           `json:"annotations,omitempty"`
-	Description *string                `json:"description,omitempty"`
-	MIMEType    *string                `json:"mimeType,omitempty"`
-	Name        string                 `json:"name"`
-	Size        *int                   `json:"size,omitempty"`
-	Title       *string                `json:"title,omitempty"`
-	URI         string                 `json:"uri"`
+	Meta        map[string]any `json:"_meta,omitempty"`
+	Annotations *Annotations   `json:"annotations,omitempty"`
+	Description *string        `json:"description,omitempty"`
+	MIMEType    *string        `json:"mimeType,omitempty"`
+	Name        string         `json:"name"`
+	Size        *int           `json:"size,omitempty"`
+	Title       *string        `json:"title,omitempty"`
+	URI         string         `json:"uri"`
 }
 
 // The contents of a specific resource or sub-resource.
 type ResourceContents struct {
-	Meta     map[string]interface{} `json:"_meta,omitempty"`
-	MIMEType *string                `json:"mimeType,omitempty"`
-	URI      string                 `json:"uri"`
+	Meta     map[string]any `json:"_meta,omitempty"`
+	MIMEType *string        `json:"mimeType,omitempty"`
+	URI      string         `json:"uri"`
 }
 
 // A resource that the server is capable of reading, included in a prompt or tool call result.
 //
 // Note: resource links returned by tools are not guaranteed to appear in the results of `resources/list` requests.
 type ResourceLink struct {
-	Meta        map[string]interface{} `json:"_meta,omitempty"`
-	Annotations *Annotations           `json:"annotations,omitempty"`
-	Description *string                `json:"description,omitempty"`
-	MIMEType    *string                `json:"mimeType,omitempty"`
-	Name        string                 `json:"name"`
-	Size        *int                   `json:"size,omitempty"`
-	Title       *string                `json:"title,omitempty"`
-	Type        string                 `json:"type"`
-	URI         string                 `json:"uri"`
+	Meta        map[string]any `json:"_meta,omitempty"`
+	Annotations *Annotations   `json:"annotations,omitempty"`
+	Description *string        `json:"description,omitempty"`
+	MIMEType    *string        `json:"mimeType,omitempty"`
+	Name        string         `json:"name"`
+	Size        *int           `json:"size,omitempty"`
+	Title       *string        `json:"title,omitempty"`
+	Type        string         `json:"type"`
+	URI         string         `json:"uri"`
 }
 
 // An optional notification from the server to the client, informing it that the list of resources it can read from has changed. This may be issued by servers without any previous subscription from the client.
 type ResourceListChangedNotification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // A template description for resources available on the server.
 type ResourceTemplate struct {
-	Meta        map[string]interface{} `json:"_meta,omitempty"`
-	Annotations *Annotations           `json:"annotations,omitempty"`
-	Description *string                `json:"description,omitempty"`
-	MIMEType    *string                `json:"mimeType,omitempty"`
-	Name        string                 `json:"name"`
-	Title       *string                `json:"title,omitempty"`
-	URITemplate string                 `json:"uriTemplate"`
+	Meta        map[string]any `json:"_meta,omitempty"`
+	Annotations *Annotations   `json:"annotations,omitempty"`
+	Description *string        `json:"description,omitempty"`
+	MIMEType    *string        `json:"mimeType,omitempty"`
+	Name        string         `json:"name"`
+	Title       *string        `json:"title,omitempty"`
+	URITemplate string         `json:"uriTemplate"`
 }
 
 // A reference to a resource or resource template definition.
@@ -520,55 +520,55 @@ type ResourceTemplateReference struct {
 
 // A notification from the server to the client, informing it that a resource has changed and may need to be read again. This should only be sent if the client previously sent a resources/subscribe request.
 type ResourceUpdatedNotification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 type Result struct {
-	Meta map[string]interface{} `json:"_meta,omitempty"`
+	Meta map[string]any `json:"_meta,omitempty"`
 }
 
 // Represents a root directory or file that the server can operate on.
 type Root struct {
-	Meta map[string]interface{} `json:"_meta,omitempty"`
-	Name *string                `json:"name,omitempty"`
-	URI  string                 `json:"uri"`
+	Meta map[string]any `json:"_meta,omitempty"`
+	Name *string        `json:"name,omitempty"`
+	URI  string         `json:"uri"`
 }
 
 // A notification from the client to the server, informing it that the list of roots has changed.
 // This notification should be sent whenever the client adds, removes, or modifies any root.
 // The server should then request an updated list of roots using the ListRootsRequest.
 type RootsListChangedNotification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // Describes a message issued to or received from an LLM API.
 type SamplingMessage struct {
-	Content interface{} `json:"content"`
-	Role    Role        `json:"role"`
+	Content any  `json:"content"`
+	Role    Role `json:"role"`
 }
 
 // Capabilities that a server may support. Known capabilities are defined here, in this schema, but this is not a closed set: any server can define its own, additional capabilities.
 type ServerCapabilities struct {
-	Completions  map[string]interface{}            `json:"completions,omitempty"`
-	Experimental map[string]map[string]interface{} `json:"experimental,omitempty"`
-	Logging      map[string]interface{}            `json:"logging,omitempty"`
-	Prompts      map[string]interface{}            `json:"prompts,omitempty"`
-	Resources    map[string]interface{}            `json:"resources,omitempty"`
-	Tools        map[string]interface{}            `json:"tools,omitempty"`
+	Completions  map[string]any            `json:"completions,omitempty"`
+	Experimental map[string]map[string]any `json:"experimental,omitempty"`
+	Logging      map[string]any            `json:"logging,omitempty"`
+	Prompts      map[string]any            `json:"prompts,omitempty"`
+	Resources    map[string]any            `json:"resources,omitempty"`
+	Tools        map[string]any            `json:"tools,omitempty"`
 }
 
-type ServerNotification interface{}
+type ServerNotification any
 
-type ServerRequest interface{}
+type ServerRequest any
 
-type ServerResult interface{}
+type ServerResult any
 
 // A request from the client to the server, to enable or adjust logging.
 type SetLevelRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 type StringSchema struct {
@@ -582,34 +582,34 @@ type StringSchema struct {
 
 // Sent from the client to request resources/updated notifications from the server whenever a particular resource changes.
 type SubscribeRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
 
 // Text provided to or from an LLM.
 type TextContent struct {
-	Meta        map[string]interface{} `json:"_meta,omitempty"`
-	Annotations *Annotations           `json:"annotations,omitempty"`
-	Text        string                 `json:"text"`
-	Type        string                 `json:"type"`
+	Meta        map[string]any `json:"_meta,omitempty"`
+	Annotations *Annotations   `json:"annotations,omitempty"`
+	Text        string         `json:"text"`
+	Type        string         `json:"type"`
 }
 
 type TextResourceContents struct {
-	Meta     map[string]interface{} `json:"_meta,omitempty"`
-	MIMEType *string                `json:"mimeType,omitempty"`
-	Text     string                 `json:"text"`
-	URI      string                 `json:"uri"`
+	Meta     map[string]any `json:"_meta,omitempty"`
+	MIMEType *string        `json:"mimeType,omitempty"`
+	Text     string         `json:"text"`
+	URI      string         `json:"uri"`
 }
 
 // Definition for a tool the client can call.
 type Tool struct {
-	Meta         map[string]interface{} `json:"_meta,omitempty"`
-	Annotations  *ToolAnnotations       `json:"annotations,omitempty"`
-	Description  *string                `json:"description,omitempty"`
-	InputSchema  map[string]interface{} `json:"inputSchema"`
-	Name         string                 `json:"name"`
-	OutputSchema map[string]interface{} `json:"outputSchema,omitempty"`
-	Title        *string                `json:"title,omitempty"`
+	Meta         map[string]any   `json:"_meta,omitempty"`
+	Annotations  *ToolAnnotations `json:"annotations,omitempty"`
+	Description  *string          `json:"description,omitempty"`
+	InputSchema  map[string]any   `json:"inputSchema"`
+	Name         string           `json:"name"`
+	OutputSchema map[string]any   `json:"outputSchema,omitempty"`
+	Title        *string          `json:"title,omitempty"`
 }
 
 // Additional properties describing a Tool to clients.
@@ -630,12 +630,12 @@ type ToolAnnotations struct {
 
 // An optional notification from the server to the client, informing it that the list of tools it offers has changed. This may be issued by servers without any previous subscription from the client.
 type ToolListChangedNotification struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params,omitempty"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params,omitempty"`
 }
 
 // Sent from the client to request cancellation of resources/updated notifications from the server. This should follow a previous resources/subscribe request.
 type UnsubscribeRequest struct {
-	Method string                 `json:"method"`
-	Params map[string]interface{} `json:"params"`
+	Method string         `json:"method"`
+	Params map[string]any `json:"params"`
 }
