@@ -63,9 +63,6 @@ type Config struct {
 	{{- else if eq $name "mcp" }}
 	// MCP settings
 	MCP *MCPConfig ` + "`env:\", prefix=MCP_\" description:\"MCP configuration\"`" + `
-	{{- else if eq $name "a2a" }}
-	// A2A settings
-	A2A *A2AConfig ` + "`env:\", prefix=A2A_\" description:\"A2A configuration\"`" + `
 	{{- else if eq $name "auth" }}
 	// Authentication settings
 	Auth *AuthConfig ` + "`env:\", prefix=AUTH_\" description:\"Authentication configuration\"`" + `
@@ -99,14 +96,6 @@ type TelemetryConfig struct {
 type MCPConfig struct {
 	{{- range $field := $section.Settings }}
 	{{ pascalCase (trimPrefix $field.Env "MCP_") }} {{ $field.Type }} ` + "`env:\"{{ trimPrefix $field.Env \"MCP_\" }}{{if $field.Default}}, default={{$field.Default}}{{end}}\" description:\"{{$field.Description}}\"`" + `
-	{{- end }}
-}
-{{- else if eq $name "a2a" }}
-
-// A2A configuration
-type A2AConfig struct {
-	{{- range $field := $section.Settings }}
-	{{ pascalCase (trimPrefix $field.Env "A2A_") }} {{ $field.Type }} ` + "`env:\"{{ trimPrefix $field.Env \"A2A_\" }}{{if $field.Default}}, default={{$field.Default}}{{end}}\" description:\"{{$field.Description}}\"`" + `
 	{{- end }}
 }
 {{- else if eq $name "auth" }}
@@ -178,13 +167,12 @@ func (cfg *Config) Load(lookuper envconfig.Lookuper) (Config, error) {
 func (cfg *Config) String() string {
     return fmt.Sprintf(
         "Config{ApplicationName:%s, Version:%s Environment:%s, Telemetry:%+v, "+
-            "MCP:%+v, A2A:%+v, Auth:%+v, Server:%+v, Client:%+v, Providers:%+v}",
+            "MCP:%+v, Auth:%+v, Server:%+v, Client:%+v, Providers:%+v}",
         APPLICATION_NAME,
         VERSION,
         cfg.Environment,
         cfg.Telemetry,
         cfg.MCP,
-        cfg.A2A,
         cfg.Auth,
         cfg.Server,
         cfg.Client,
